test: Isolate the page-size reset and split the URL-adapter suite
Second round of CodeRabbit CLI findings, both on the test file.

The page-size reset test passed a freshly built rows() array on rerender, so
`data` identity changed alongside pageSize and the test could not tell its own
subject from an ordinary auto-reset on a data change. One array is now reused
across both renders, leaving pageSize as the only variable. Re-mutated to
confirm it still fails when the reset is removed, so it now fails for the
reason it claims.

The tableUrlParams and sortParser assertions moved to their own suite.
#377's boundary is that useTableState.ts imports nothing from
tableUrlStore.ts, and a single suite spanning both quietly coupled the two
layers that decision exists to keep apart.
